Work Cited Page
I am adding a work cited page because I don't consider this to be my original work. I wanted to complete the assignment and understand the thought process of how to get to the end result. During my debugging process I came across this website. I didn't copy it and moved on, I spend a good amount of time understanding each line of code. I understand if I don't get a grade for this assignment.